ASoC: wcd-mbhc: Fix headphone detection
Currently, headphones are detected as a headset, which causes a problem
when the user wants to use a microphone. In this situation, the on-board
microphone is erroneously disabled and the system attempts to enable the
microphone on the attached peripheral, even though there isn't any
microphone on a regular pair of headphones, causing no microphone to be
available for input.

Clean up the button-interrupt disable logic to prevent interference with
plug correction, and fix the plug correction logic so that headphones are
detected properly.

Signed-off-by: Sultanxda <sultanxda@gmail.com>